backend/requirements.txt
# ─────────────────────────────────────────────
# mycelo — backend requirements
# Install: pip install -r requirements.txt
# ─────────────────────────────────────────────

# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1

# NLP — data_to_graph pipeline
gliner==0.2.7
transformers==4.40.0
sentence-transformers==2.7.0
spacy==3.7.4
torch==2.3.0
python-dateparser==1.2.0

# Database — graph
gqlalchemy==1.6

# Database — embeddings
sqlalchemy==2.0.48
psycopg[binary]==3.2.1
pgvector==0.3.5

# Utilities
numpy==1.26.4